requirements.txt
numpy==1.21.4
pandas==1.3.5
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
discord-webhook==0.14.0
mysql-connector-python==8.0.23
tweepy